Separate restore code from the rest
